Fix ProjectBuildStatusAssetStrategy with tests
